\name{BRIGHT-package}
\alias{BRIGHT-package}
\alias{BRIGHT}
\docType{package}
\title{\packageTitle{BRIGHT}}
\description{\packageDescription{BRIGHT}}
\section{Package Content}{\packageIndices{BRIGHT}}
\author{\packageAuthor{BRIGHT}}
\section{Maintainer}{\packageMaintainer{BRIGHT}}
\keyword{package}
