% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F}
\alias{F}
\title{Firm-thresholding}
\usage{
F(z, l1, l2, gamma)
}
\arguments{
\item{z}{a numeric scalar to be Firm-thresholded}

\item{l1}{a numeric scalar for the first Firm-thresholding parameter}

\item{l2}{a numeric scalar for the second Firm-thresholding parameter}

\item{gamma}{a numeric scalar for the third Firm-thresholding parameter}
}
\value{
F(z,l1,l2)
}
\description{
Calculate the Firm-thresholding result
}
