% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Ff}
\alias{Ff}
\title{Firm-thresholding}
\usage{
Ff(z, l1, l2, gamma)
}
\arguments{
\item{z}{a numeric scalar to be firm-thresholded.}

\item{l1}{a numeric scalar for the first firm-thresholding parameter.}

\item{l2}{a numeric scalar for the second firm-thresholding parameter.}

\item{gamma}{a numeric scalar for the third firm-thresholding parameter.}
}
\value{
A firm-thresholded scalar of z.
}
\description{
Calculate the Firm-thresholding result corresponding to the MCP penalty.
}
\examples{
Ff(5,1,1,1)
Ff(0.5,1,1,1)
}
