% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdfit_gaussian}
\alias{gdfit_gaussian}
\title{BRIGHT estimation procedure}
\usage{
gdfit_gaussian(
  XtY,
  tilde_beta,
  X,
  lambda,
  K1,
  m,
  blk,
  K0,
  penalty,
  tau,
  eta,
  alpha,
  gamma,
  eps,
  max_iter,
  dfmax,
  gmax,
  user
)
}
\arguments{
\item{XtY}{a numeric verticle vector of the marginal correlation between genotype and outcome, can be recovered from GWAS summary statistics by function p2cor().}

\item{tilde_beta}{a numeric verticle vector of the prior information; in the case where prior information is a subspace of XtY, set the complement set to be zero.}

\item{X}{a numeric matrix of reference genotypes, default is from 1000 genome project.}

\item{lambda}{a numeric verticle vector of the LASSO penalty weights; it must be sorted with a decreasing order.}

\item{K1}{a numeric verticle vector of the grouping of SNPs for group penalties; SNPs within K1[i,i+1] are grouped together.}

\item{m}{a numeric verticle vector of the multiplicative factor to adjust for the number of SNPs in each group, the default is the square root of number of elements in each group.}

\item{blk}{a numeric verticle vector indicating the grouping of SNPs in each block, can be generated through LD() function.}

\item{K0}{a integer scalar for the number of SNPs/covariates that is not penalized.}

\item{penalty}{a integer scalar for chosing the penalties; 1 corresponds to LASSO, 2 corresponds to MCP, 3 corresponds to SCAD.}

\item{tau}{a numeric scalar represents the element-wise soft-thresholding parameter for generating the LD.}

\item{eta}{a numeric scalar of the weights for incorporating prior information.}

\item{alpha}{a numeric scalar of the ridge penalty weights; alpha=1 corresponds to no ridge penalty.}

\item{gamma}{a numeric scalar of the third parameter in firm and SCAD thresholding functions.}

\item{eps}{a numeric scalar for the convergence criteria.}

\item{max_iter}{an integer scalar for the maximum iterations before the algorithm stops.}

\item{dfmax}{an integer scalar for the maximum number of SNPs been selected before the algorithm stops.}

\item{gmax}{an integer scalar for the maximum number of groups been selected before the algorithm stops.}

\item{user}{a logical scalar for indicating whether lambda is user specified; if user=TRUE, then the iteration will start from the largest lambda, otherwise the iteration will start from the second largest lambda.}
}
\value{
Write a list of results; beta, the coefficient estimate from BRIGHT;

iter, the number of total iterations needed for the model to converge with each lambda;

df total degree of freedom of the converged model with each lambda;

dev, the approximated deviance associated with each lambda.
}
\description{
Gradient-descent algorithm for optimizing the BRIGHT estimation procedure
}
