% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LD}
\alias{LD}
\title{LD Block Structured SNPs grouping}
\usage{
LD(chr, pos, LDB)
}
\arguments{
\item{chr}{a numeric verticle vector of the chromosomes each SNP belong to.}

\item{pos}{a numeric verticle vector of the position in base pair (BP) of each SNP.}

\item{LDB}{a numeric matrix for the block structure information directly read in from Berisa's bed file.}
}
\value{
A numeric verticle vector, representing the block structured grouping of the current SNP list; SNP with (LD[i],LD[i+1]) are to be within one group.
}
\description{
Calculate the LD grouping of input SNPs based on existing block structures (Berisa et. al.). The chr and pos must be matched and sorted.
}
\examples{
chr=t(t(c(1,1,1,2,2,3,3)))
pos=t(t(c(10,20,30,1,20,9,20)))
LDB=data.frame(chr=c(1,1,1,2,2,2,3,3,3),start=c(1,15,40,1,30,60,1,50,100),stop=c(15,40,80,30,60,100,50,100,150))
LDB=as.matrix(LDB)
LD(chr,pos,LDB)
}
