% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaxLambda}
\alias{MaxLambda}
\title{Maximum Lambda selection}
\usage{
MaxLambda(XtY, tilde_beta, X, K1, m, blk, K0, tau, eta, alpha, eps, max_iter)
}
\arguments{
\item{XtY}{a numeric verticle vector of the marginal correlation between genotype and outcome, can be recovered from GWAS summary statistics by function p2cor().}

\item{tilde_beta}{a numeric verticle vector of the prior information; in the case where prior information is a subspace of XtY, set the complement set to be zero.}

\item{X}{a numeric matrix of reference genotypes, default is from 1000 genome project.}

\item{K1}{a numeric verticle vector of the grouping of SNPs for group penalties; SNPs within K1[i,i+1] are grouped together.}

\item{m}{a numeric verticle vector of the multiplicative factor to adjust for the number of SNPs in each group, the default is the square root of number of elements in each group.}

\item{blk}{a numeric verticle vector indicating the grouping of SNPs in each block, can be generated through LD() function.}

\item{K0}{a integer scalar for the number of SNPs/covariates that is not penalized.}

\item{tau}{a numeric scalar represents the element-wise soft-thresholding parameter for generating the LD.}

\item{eta}{a numeric scalar of the weights for incorporating prior information.}

\item{alpha}{a numeric scalar of the ridge penalty weights; alpha=1 corresponds to no ridge penalty.}

\item{eps}{a numeric scalar for the convergence criteria.}

\item{max_iter}{an integer scalar for the maximum iterations before the algorithm stops.}

\item{lambda}{a numeric verticle vector of the LASSO penalty weights; it must be sorted with a decreasing order.}
}
\description{
Calculate the maximum lambda, which corresponding to penalizing all coefficients to be zero
}
