% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Norm}
\alias{Norm}
\title{Second Norm}
\usage{
Norm(x, p)
}
\arguments{
\item{x}{a numeric column vector to calculate the second norm.}

\item{p}{an integer represents the dimension of x.}
}
\value{
A scalar of the computed second norm of x.
}
\description{
Calculate second norm (Euclidean distance) of a vertical vector.
}
\examples{
x <- t(t(rnorm(100)))
Norm(x,length(x))
}
