% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2cor.R
\name{p2cor}
\alias{p2cor}
\title{GWAS to correlation}
\usage{
p2cor(p, n, sign = rep(1, length(p)))
}
\arguments{
\item{p}{a numeric vector of the p values from marginal SLR.}

\item{n}{a numeric vector of the sample sizes.}

\item{sign}{a numeric vector of the signs of coefficient estimates.}
}
\value{
Recovered XtY/n from the GWAS summary.
}
\description{
Recover marginal correlation information from GWAS summary statistics.
}
\examples{
p = c(0.01,0.015)
n = c(100,1000)
sign = c(-1,1)
p2cor(p, n, sign)
}
