% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRIGHT.R
\name{Preprocess}
\alias{Preprocess}
\title{Preprocess}
\usage{
Preprocess(chr, BP, A1, A2, p, n, sign, ref = "EUR", group = 1:p, m = NA)
}
\arguments{
\item{chr}{a numeric vector of the chromesomes of each SNP in GWAS.}

\item{BP}{a numeric vector of the position in base pair of each SNP in GWAS.}

\item{A1}{a string vector of the effect allele of each SNP in GWAS.}

\item{A2}{a string vector of the reference allele of each SNP in GWAS.}

\item{p}{a numeric vector of the p values from marginal t-test.}

\item{n}{a numeric vector of the sample sizes.}

\item{sign}{a numeric vector of the signs of coefficient estimates.}

\item{ref}{a string of the GWAS population; EUR, European ancestry; AFR, African ancestry; AMR, admixed American; EAS, East Asian; SAS, South Asian.}

\item{group}{a vector describing the grouping of the coefficients.  For
greatest efficiency and least ambiguity (see details), it is best if
\code{group} is a factor or vector of consecutive integers, although
unordered groups and character vectors are also allowed.  If there are
coefficients to be included in the model without being penalized, assign
them to group 0 (or \code{"0"}).}

\item{m}{a vector of values representing multiplicative
factors by which each group's penalty is to be multiplied.  Often, this is a
function (such as the square root) of the number of predictors in each
group.  The default is to use the square root of group size for the group
selection methods, and a vector of 1's (i.e., no adjustment for group size)
for bi-level selection.}
}
\value{
A list of preprocessed XtY and reference genotype matrix.
}
\description{
Preprocess the GWAS and the 1000 genome reference.
}
