% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC}
\alias{QC}
\title{Quality control}
\usage{
QC(ss, ind, KG, lab)
}
\arguments{
\item{ss}{a data frame of the summary statistics, which needs to match the ind indicator.}

\item{ind}{a character scalar indicator of the type of ss, it can only takes value of "GWAS", "Corr", or "Coef".}

\item{KG}{a data frame of the markers information from reference.}

\item{lab}{a character scalar of the data name being processed. Usually either "target" or "prior" as input.}
}
\value{
The new cleaned summary statistics stored as a data frame.
}
\description{
Convert GWAS summary statistics to the marginal correlations. Remove summary statistics markers, which are not in the reference or have NA values. Adjusting the summary statistics sign based on reference alleles.
}
