% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sigma_LD}
\alias{Sigma_LD}
\title{Block Structured LD Calculation}
\usage{
Sigma_LD(X, Sig, blk, tau, P, N)
}
\arguments{
\item{X}{a numeric matrix of the standardized genotypes outputted through newXG() function.}

\item{Sig}{a numeric sparse matrix pointer, indicating the memory location where the output sparse LD matrix should be stored.}

\item{blk}{a numeric verticle vector indicating the grouping of SNPs in each block, can be generated through LD() function.}

\item{tau}{a numeric scalar represents the element-wise soft-thresholding parameter for generating the LD.}

\item{P}{an integer scalar for the column dimension (genotype dimension) of X.}

\item{N}{an integer scalar for the row dimension (sample size) of X.}
}
\value{
Write the calculated sparse LD matrix to the memory location provided in Sig.
}
\description{
Calculate the LD based on existing block structures.
}
