% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm}
\alias{norm}
\title{Second norm}
\usage{
norm(x, p)
}
\arguments{
\item{x}{a numeric column vector to calculate the second norm}

\item{p}{a integer represents the dimension of x}
}
\value{
||x||_2
}
\description{
Calculate second norm of a vertical vector
}
