% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S}
\alias{S}
\title{Soft-thresholding}
\usage{
S(z, l)
}
\arguments{
\item{z}{a numeric scalar to be soft-thresholded.}

\item{l}{a numeric scalar for the soft-thresholding parameter.}
}
\value{
If z <= l, 0 is outputted.

If z > l, z-l is outputted.
}
\description{
Calculate the soft-thresholding result corresponding to the LASSO penalty.
}
\examples{
S(5,1)
S(0.5,1)
}
